+++ src/schemas/finnance/constants/index.ts
@@ -1,5 +1,7 @@
 import { ACCOUNT_TYPES } from "./accountType.js";
 import { IBAN } from "./iban.js";
 import { MONEY_INFO } from "./moneyInfo.js";
+import { SUBSCRIPTION_PLAN } from "./subscription.js";
+import { TRANSACTION_TYPE } from "./transactionType.js";
 
-export { ACCOUNT_TYPES, IBAN, MONEY_INFO };
+export { ACCOUNT_TYPES, IBAN, MONEY_INFO, SUBSCRIPTION_PLAN, TRANSACTION_TYPE };