+++ src/schemas/internet/constants/browser.ts
@@ -1,0 +1,13 @@
+export const BROWSERS = [
+  "Chrome",
+  "Edge",
+  "Firefox",
+  "Internet Explorer",
+  "Safari",
+  "Opera",
+  "Yandex",
+  "Chromium",
+  "Vivaldi",
+  "Baidu",
+  "Brave",
+];