+++ src/schemas/internet/constants/protocol.ts
@@ -1,0 +1,13 @@
+export const PROTOCOL = [
+  "https",
+  "http",
+  "tcp",
+  "udp",
+  "ip",
+  "pop",
+  "smtp",
+  "dhcp",
+  "l2tp",
+  "ftp",
+  "imap",
+];